Removed looping (caused bug). Now it is getting duration and right notes when recording. Still don't get playback.

+++ ucimusic147-6/ucimusic147-6/MUS147Sequencer.h
@@ -17,7 +17,7 @@
     BOOL recording;
     UInt32 numEvents;
     Float64 loopLength;
-    UInt32 noteOn[128];
+    Float64 noteOn[128];
 }
 
 @property (readwrite) Float64 scoreTime;
